+++ src/integration-test/java/org/openlmis/auth/web/ApiKeyControllerIntegrationTest.java
@@ -100,10 +100,7 @@ public class ApiKeyControllerIntegrationTest extends BaseWebIntegrationTest {
   private RightDto right = new DummyRightDto();
 
   @Before
-  @Override
   public void setUp() {
-    super.setUp();
-
     client = new Client();
     key = new ApiKeyDataBuilder().build();