+++ src/test/java/org/openlmis/auth/dto/LocalizedMessageDtoTest.java
@@ -13,20 +13,27 @@
  * http://www.gnu.org/licenses.  For additional information contact info@OpenLMIS.org. 
  */
 
-package org.openlmis.auth.dto.referencedata;
+package org.openlmis.auth.dto;
 
-import com.fasterxml.jackson.annotation.JsonInclude;
-import lombok.EqualsAndHashCode;
-import lombok.Getter;
-import lombok.Setter;
+import nl.jqno.equalsverifier.EqualsVerifier;
+import nl.jqno.equalsverifier.Warning;
+import org.junit.Test;
+import org.openlmis.auth.ToStringTestUtils;
 
-import java.util.UUID;
+public class LocalizedMessageDtoTest {
 
-@JsonInclude(JsonInclude.Include.NON_NULL)
-@EqualsAndHashCode
-public class BaseDto {
-  
-  @Getter
-  @Setter
-  private UUID id;
+  @Test
+  public void equalsContract() {
+    EqualsVerifier
+        .forClass(LocalizedMessageDto.class)
+        .suppress(Warning.NONFINAL_FIELDS) // dto can't contain final fields
+        .verify();
+  }
+
+  @Test
+  public void shouldImplementToString() {
+    LocalizedMessageDto account = new LocalizedMessageDto("key", "value");
+    ToStringTestUtils.verify(LocalizedMessageDto.class, account);
+  }
+
 }